+++ CHANGELOG.md
@@ -6,9 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v0.2.0] - 2019-07-16
+
 ### Changed
 - Thing.publish() removed. Self links are not used at all (and not recommended for static catalogs)
-- Thing.root() and Thing.parent() functions now return `None` if no root or parent (rather than an empty list). If more than one root or parent then an error will now be thrown.
+- Thing.root() returns Thing if no root (rather than an empty list). If more than one root an error will now be thrown.
+- Thing.parent() functions now return `None` if no parent (rather than an empty list). If more than one parent then an error will now be thrown.
 - Internal JSON data now stored in variable called `_data` rather than `data`
 
 ## [v0.1.3] - 2019-05-04
@@ -45,6 +48,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 Initial Release
 
 [Unreleased]: https://github.com/sat-utils/sat-stac/compare/master...develop
+[v0.2.0]: https://github.com/sat-utils/sat-stac/compare/0.1.3...v0.2.0
 [v0.1.3]: https://github.com/sat-utils/sat-stac/compare/0.1.2...v0.1.3
 [v0.1.2]: https://github.com/sat-utils/sat-stac/compare/0.1.1...v0.1.2
 [v0.1.1]: https://github.com/sat-utils/sat-stac/compare/0.1.0...v0.1.1